# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Todo API
  description: A simple Todo API built with Rust and Axum
  license:
    name: ''
  version: 1.0.0
paths:
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login user
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh access token
      operationId: refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid refresh token
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register a new user
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
        '409':
          description: Email already registered
  /api/v1/todos:
    get:
      tags:
      - todos
      summary: List all todos for authenticated user
      operationId: list_todos
      parameters:
      - name: page
        in: query
        description: 'Page number (default: 1)'
        required: false
        schema:
          type: integer
          format: int64
      - name: per_page
        in: query
        description: 'Items per page (default: 20, max: 100)'
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: List of todos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoListResponse'
        '401':
          description: Unauthorized
      security:
      - bearer_auth: []
    post:
      tags:
      - todos
      summary: Create a new todo
      operationId: create_todo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTodoRequest'
        required: true
      responses:
        '201':
          description: Todo created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
      security:
      - bearer_auth: []
  /api/v1/todos/{id}:
    get:
      tags:
      - todos
      summary: Get a specific todo
      operationId: get_todo
      parameters:
      - name: id
        in: path
        description: Todo ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '401':
          description: Unauthorized
        '404':
          description: Todo not found
      security:
      - bearer_auth: []
    put:
      tags:
      - todos
      summary: Update a todo
      operationId: update_todo
      parameters:
      - name: id
        in: path
        description: Todo ID
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTodoRequest'
        required: true
      responses:
        '200':
          description: Todo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResponse'
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '404':
          description: Todo not found
      security:
      - bearer_auth: []
    delete:
      tags:
      - todos
      summary: Delete a todo
      operationId: delete_todo
      parameters:
      - name: id
        in: path
        description: Todo ID
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Todo deleted
        '401':
          description: Unauthorized
        '404':
          description: Todo not found
      security:
      - bearer_auth: []
components:
  schemas:
    AuthResponse:
      type: object
      required:
      - access_token
      - refresh_token
      - token_type
      - expires_in
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
          format: int64
        refresh_token:
          type: string
        token_type:
          type: string
    CreateTodoRequest:
      type: object
      required:
      - title
      properties:
        description:
          type:
          - string
          - 'null'
        title:
          type: string
    LoginRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
        password:
          type: string
    RefreshRequest:
      type: object
      required:
      - refresh_token
      properties:
        refresh_token:
          type: string
    RegisterRequest:
      type: object
      required:
      - email
      - password
      properties:
        email:
          type: string
        password:
          type: string
    Todo:
      type: object
      required:
      - id
      - user_id
      - title
      - completed
      - created_at
      - updated_at
      properties:
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        description:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        title:
          type: string
        updated_at:
          type: string
          format: date-time
        user_id:
          type: string
          format: uuid
    TodoListResponse:
      type: object
      required:
      - todos
      - total
      - page
      - per_page
      properties:
        page:
          type: integer
          format: int64
        per_page:
          type: integer
          format: int64
        todos:
          type: array
          items:
            $ref: '#/components/schemas/TodoResponse'
        total:
          type: integer
          format: int64
    TodoResponse:
      type: object
      required:
      - id
      - title
      - completed
      - created_at
      - updated_at
      properties:
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        description:
          type:
          - string
          - 'null'
        id:
          type: string
          format: uuid
        title:
          type: string
        updated_at:
          type: string
          format: date-time
    UpdateTodoRequest:
      type: object
      properties:
        completed:
          type:
          - boolean
          - 'null'
        description:
          type:
          - string
          - 'null'
        title:
          type:
          - string
          - 'null'
    User:
      type: object
      required:
      - id
      - email
      - created_at
      - updated_at
      properties:
        created_at:
          type: string
          format: date-time
        email:
          type: string
        id:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
    UserResponse:
      type: object
      required:
      - id
      - email
      properties:
        email:
          type: string
        id:
          type: string
          format: uuid
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
tags:
- name: auth
  description: Authentication API
- name: todos
  description: Todo management API